requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
PyMySQL==1.1.0
mysqlclient==2.2.0
akshare==1.11.80
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
django-extensions==3.2.3
django-filter==23.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
scikit-learn==1.3.2
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
psutil==5.9.6
APScheduler==3.10.4
openai==1.3.7
minio==7.2.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7